backend/requirements.txt
Django
djangorestframework
psycopg2-binary           
matplotlib
tensorflow
keras
pillow
django-cors-headers
djangorestframework-simplejwt
django-filter